requirements.txt
flask
Flask-SQLAlchemy
Flask-Cors
Flask_WTF
Flask_Login
Flask_User
Flask_Migrate
Flask_BabelEx
Jinja2
python-dotenv
requests
authlib>=0.14.1
six
pycryptodome==3.3.1
python-jose-cryptodome
Werkzeug
shortid
maya
Flask_Admin
gunicorn
psycopg2
alembic